+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # Cynosure `@VERSION@` `1.20.1` `Fabric & Forge`
 
 ### Changelog:
-- Fixed Crash with GlobalStorage
+- Fixed keybind error on Fabric